BUG: Fix the method to add parachutes to the rocket
The code changes in `nb_builder.py` add a new method `add_parachutes_to_rocket` to the `NotebookBuilder` class. This method adds parachutes to the rocket by updating the `rocket.parachutes` attribute with the values from the `parachutes` dictionary.

This commit message follows the established convention of using a prefix to indicate the type of change (`feat` for a new feature) and provides a clear and concise description of the purpose of the code changes.